requirements.txt
streamlit
chromadb
google-generativeai
sentence-transformers
python-dotenv